restructure after adding starting-page
