+++ components/RandomCover/index.js
@@ -2,10 +2,12 @@ import Link from "next/link";
 import Image from "next/image";
 
 export default function RandomCover({ artworks }) {
-  const randomArtwork = artworks[Math.floor(Math.random() * artworks.length)];
+  const randomArtwork =
+    artworks.items[Math.floor(Math.random() * artworks.items.length)];
+  console.log(artworks);
   return (
     <>
-      <Link href={`/artwork-info/${id}`}>
+      <Link href={`/artwork-info/${randomArtwork.id}`}>
         <Image
           src={randomArtwork.image_thumbnail}
           alt={randomArtwork.titleText}
