+++ pages/artworks-collection/index.js
@@ -1,0 +1,20 @@
+import ArtworksList from "@/components/ArtworksList/index.js";
+
+export default function HomePage({
+  artworks,
+  onHandlePreviousPage,
+  onHandleNextPage,
+  offset,
+  rowsPerPage,
+}) {
+  return (
+    <>
+      <h1>Artworks Collection</h1>
+      <ArtworksList artworks={artworks} />
+      {offset >= rowsPerPage && (
+        <button onClick={onHandlePreviousPage}>Previous Page</button>
+      )}
+      <button onClick={onHandleNextPage}>Next Page</button>
+    </>
+  );
+}
